requirements.txt
python-telegram-bot
google-generativeai
flask
